+++ frontend/components/Customer/Header/index.js
@@ -1,0 +1,5 @@
+import HeaderSection from '../../HeaderSection';
+
+const Header = () => <HeaderSection />;
+
+export default Header;
