+++ frontend/components/Customer/ProgramNotActive/index.js
@@ -1,4 +1,4 @@
-import LoyaltyPrograms from '../../../../components/LoyaltyPrograms';
+import LoyaltyPrograms from '../../LoyaltyPrograms';
 
 const ProgramNotActive = ({ programsList }) => <LoyaltyPrograms programsList={programsList} details={false} />;
